Get the test to pass
- Split main away from index like it was originally. This should also
  make future tests easier as we can import the code as a library
  without running it.
- Remove the part of the test that doesn't work yet.
